# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The backend powering most things",
        "title": "web-api",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/ystv/web-api",
            "email": "computing@ystv.co.uk"
        },
        "license": {}
    },
    "paths": {
...
        "/v1/internal/misc/quotes": {
            "put": {
                "description": "updates a quote. Still need to provide the whole Quote object,\nweb-api will overwrite created by User ID to keep with existing record.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "misc-quotes"
                ],
                "summary": "Update quote",
                "operationId": "update-quote",
                "parameters": [
                    {
                        "description": "Quote object",
                        "name": "quote",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/misc.Quote"
                        }
                    }
                ],
                "responses": {
                    "200": {}
                }
            },
            "post": {
                "description": "creates a new quote.\nweb-api will overwrite created by User ID with the token's user ID.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "misc-quotes"
                ],
                "summary": "New quote",
                "operationId": "new-quote",
                "parameters": [
                    {
                        "description": "Quote object",
                        "name": "quote",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/misc.Quote"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Quote ID",
                        "schema": {
                            "$ref": "#/definitions/int"
                        }
                    }
                }
            }
        },
        "/v1/internal/misc/quotes/{amount}/{page}": {
            "get": {
                "description": "Lists quotes by pagination.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "misc-quotes"
                ],
                "summary": "List quotes",
                "operationId": "get-quotes",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Amount",
                        "name": "amount",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/misc.QuotePage"
                            }
                        }
                    }
                }
            }
        },
        "/v1/internal/misc/quotes/{quoteid}": {
            "delete": {
                "description": "deletes a quote by ID.",
                "tags": [
                    "misc-quotes"
                ],
                "summary": "Delete quote",
                "operationId": "delete-quote",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Quote ID",
                        "name": "quoteid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {}
                }
            }
        },
...
        "/v1/internal/misc/webcams/{cameraid}": {
            "get": {
                "description": "Reverse proxies the selected webcam returns the jpeg feed as a result.",
                "tags": [
                    "misc-webcams"
                ],
                "summary": "Get webcam",
                "operationId": "get-webcam",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Camera ID",
                        "name": "cameraid",
                        "in": "path",
                        "required": true
                    }
                ]
            }
        },
...
        "/v1/internal/people/user/{userid}": {
            "get": {
                "description": "Get a basic user object by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "people-user"
                ],
                "summary": "Get a user by ID",
                "operationId": "get-user-id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/people.User"
                        }
                    }
                }
            }
        },
        "/v1/internal/people/user/{userid}/full": {
            "get": {
                "description": "Get a complete user object by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "people-user"
                ],
                "summary": "Get a full user by ID",
                "operationId": "get-user-id-full",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/people.User"
                        }
                    }
                }
            }
        },
...
        "/v1/internal/people/users/{roleid}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "people-users"
                ],
                "summary": "List all users of a given role",
                "operationId": "get-people-users-roles-members",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "roleid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/misc.List"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "misc.List": {
            "type": "object",
            "properties": {
                "alias": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "isSubscribed": {
                    "type": "boolean"
                },
                "listID": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "permissionID": {
                    "type": "string"
                },
                "subscribers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/misc.Subscriber"
                    }
                }
            }
        },
        "misc.Quote": {
            "type": "object",
            "properties": {
                "createdBy": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "quote": {
                    "type": "string"
                }
            }
        },
        "misc.QuotePage": {
            "type": "object",
            "properties": {
                "lastPageIndex": {
                    "type": "integer"
                },
                "quotes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/misc.Quote"
                    }
                }
            }
        },
        "misc.Subscriber": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastName": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/people.Permission"
                    }
                },
                "subscribeID": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "people.Permission": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "people.User": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastName": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/people.Permission"
                    }
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    }
}